feat:interface 주입 생성
